Add docs page
This includes:
  - New linting configs
  - Addition of routing
  - temp adds doc directory
