+++ docs/asset-manifest.json
@@ -1,0 +1,39 @@
+{
+  "files": {
+    "main.css": "/static/css/main.71ca6ac6.chunk.css",
+    "main.js": "/static/js/main.20033631.chunk.js",
+    "main.js.map": "/static/js/main.20033631.chunk.js.map",
+    "runtime-main.js": "/static/js/runtime-main.c9859e0d.js",
+    "runtime-main.js.map": "/static/js/runtime-main.c9859e0d.js.map",
+    "static/css/2.0a9ec390.chunk.css": "/static/css/2.0a9ec390.chunk.css",
+    "static/js/2.9284eac2.chunk.js": "/static/js/2.9284eac2.chunk.js",
+    "static/js/2.9284eac2.chunk.js.map": "/static/js/2.9284eac2.chunk.js.map",
+    "static/js/3.20a0c265.chunk.js": "/static/js/3.20a0c265.chunk.js",
+    "static/js/3.20a0c265.chunk.js.map": "/static/js/3.20a0c265.chunk.js.map",
+    "static/css/4.141bfd39.chunk.css": "/static/css/4.141bfd39.chunk.css",
+    "static/js/4.283b7797.chunk.js": "/static/js/4.283b7797.chunk.js",
+    "static/js/4.283b7797.chunk.js.map": "/static/js/4.283b7797.chunk.js.map",
+    "static/css/5.d78a978c.chunk.css": "/static/css/5.d78a978c.chunk.css",
+    "static/js/5.23c74361.chunk.js": "/static/js/5.23c74361.chunk.js",
+    "static/js/5.23c74361.chunk.js.map": "/static/js/5.23c74361.chunk.js.map",
+    "static/css/6.c1bf81ac.chunk.css": "/static/css/6.c1bf81ac.chunk.css",
+    "static/js/6.fcd0f960.chunk.js": "/static/js/6.fcd0f960.chunk.js",
+    "static/js/6.fcd0f960.chunk.js.map": "/static/js/6.fcd0f960.chunk.js.map",
+    "static/js/7.ac09bab5.chunk.js": "/static/js/7.ac09bab5.chunk.js",
+    "static/js/7.ac09bab5.chunk.js.map": "/static/js/7.ac09bab5.chunk.js.map",
+    "index.html": "/index.html",
+    "static/css/2.0a9ec390.chunk.css.map": "/static/css/2.0a9ec390.chunk.css.map",
+    "static/css/4.141bfd39.chunk.css.map": "/static/css/4.141bfd39.chunk.css.map",
+    "static/css/5.d78a978c.chunk.css.map": "/static/css/5.d78a978c.chunk.css.map",
+    "static/css/6.c1bf81ac.chunk.css.map": "/static/css/6.c1bf81ac.chunk.css.map",
+    "static/css/main.71ca6ac6.chunk.css.map": "/static/css/main.71ca6ac6.chunk.css.map",
+    "static/js/2.9284eac2.chunk.js.LICENSE.txt": "/static/js/2.9284eac2.chunk.js.LICENSE.txt"
+  },
+  "entrypoints": [
+    "static/js/runtime-main.c9859e0d.js",
+    "static/css/2.0a9ec390.chunk.css",
+    "static/js/2.9284eac2.chunk.js",
+    "static/css/main.71ca6ac6.chunk.css",
+    "static/js/main.20033631.chunk.js"
+  ]
+}
