+++ docs/static/css/4.141bfd39.chunk.css
@@ -1,0 +1,2 @@
+div[data-sort-dropdown],div[data-sort-responsive-list]{display:none}div[data-sort-responsive-list-always-show]{display:block!important}table[data-sort-responsive-has-list-always-hide]{display:none!important}div[data-sort-responsive-list]{width:100%;padding:10px}div[data-sort-responsive-list] dt{display:block;float:left;font-weight:700;text-transform:uppercase}div[data-sort-responsive-list] dd,div[data-sort-responsive-list] dt{font-size:.8em}div[data-sort-responsive-list] dd{text-align:right;border-bottom:1px solid #ddd;min-height:1.6em;margin-bottom:.625em}div[data-sort-responsive-list] dl{border-bottom:3px solid #ddd}div[data-sort-responsive-list] dl dt:first-child{position:absolute;left:-999999px}div[data-sort-responsive-list] dl dd:first-of-type{font-weight:700;font-size:1.2em;text-align:left}div[data-sort-responsive-list] dl dd:last-of-type{margin-bottom:0}@media screen and (max-width:700px){table[data-sort-responsive-has-list]{display:none}div[data-sort-responsive-list]{display:block}div[data-sort-dropdown]{display:flex}}
+/*# sourceMappingURL=4.141bfd39.chunk.css.map */
