+++ docs/static/css/6.c1bf81ac.chunk.css
@@ -1,0 +1,2 @@
+div[data-sort-dropdown],table[data-sort-responsive] td span[data-responsive-header]{display:none}@media screen and (max-width:700px){table[data-sort-responsive]{border:0}table[data-sort-responsive] caption{font-size:1.3em;caption-side:top}table[data-sort-responsive] thead{left:-9999999px;position:absolute}table[data-sort-responsive] td span[data-responsive-header]{display:block;float:left;font-weight:700;text-transform:uppercase}table[data-sort-responsive] tr{border-bottom:3px solid #ddd;display:block;margin-bottom:.625em}table[data-sort-responsive] td,table[data-sort-responsive] th{border-bottom:1px solid #ddd;display:block;min-height:2em}table[data-sort-responsive] td{font-size:.8em;text-align:right}table[data-sort-responsive] th{font-size:1.2em}div[data-sort-dropdown]{display:flex}}
+/*# sourceMappingURL=6.c1bf81ac.chunk.css.map */
