+++ docs/static/css/main.71ca6ac6.chunk.css
@@ -1,0 +1,2 @@
+[data-sort-table-confirm]{position:absolute;left:-99999px}@media screen and (max-width:700px){[data-filter]{text-align:left}}.emptyCell:after{content:"unknown";font-style:italic;color:#6e6e6e}.sortedCellClass{background-color:rgba(255,207,0,.1803921568627451)}
+/*# sourceMappingURL=main.71ca6ac6.chunk.css.map */
