+++ docs/static/js/3.20a0c265.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonp@mydobie/sort-table"]=this["webpackJsonp@mydobie/sort-table"]||[]).push([[3],{38:function(e,t,n){"use strict";var o=n(39).DebounceInput;o.DebounceInput=o,e.exports=o},39:function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.DebounceInput=void 0;var o=i(n(1)),r=i(n(40));function i(e){return e&&e.__esModule?e:{default:e}}function u(e){return u="function"===typeof Symbol&&"symbol"===typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"===typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},u(e)}function c(e,t){if(null==e)return{};var n,o,r=function(e,t){if(null==e)return{};var n,o,r={},i=Object.keys(e);for(o=0;o<i.length;o++)n=i[o],t.indexOf(n)>=0||(r[n]=e[n]);return r}(e,t);if(Object.getOwnPropertySymbols){var i=Object.getOwnPropertySymbols(e);for(o=0;o<i.length;o++)n=i[o],t.indexOf(n)>=0||Object.prototype.propertyIsEnumerable.call(e,n)&&(r[n]=e[n])}return r}function f(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var o=Object.getOwnPropertySymbols(e);t&&(o=o.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,o)}return n}function a(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?f(Object(n),!0).forEach((function(t){d(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):f(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function l(e,t){for(var n=0;n<t.length;n++){var o=t[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}function s(e,t){return s=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e},s(e,t)}function p(e){var t=function(){if("undefined"===typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"===typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(e){return!1}}();return function(){var n,o=v(e);if(t){var r=v(this).constructor;n=Reflect.construct(o,arguments,r)}else n=o.apply(this,arguments);return y(this,n)}}function y(e,t){return!t||"object"!==u(t)&&"function"!==typeof t?b(e):t}function b(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function v(e){return v=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)},v(e)}function d(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var h=function(e){!function(e,t){if("function"!==typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&s(e,t)}(f,e);var t,n,i,u=p(f);function f(e){var t;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,f),d(b(t=u.call(this,e)),"onChange",(function(e){e.persist();var n=t.state.value,o=t.props.minLength;t.setState({value:e.target.value},(function(){var r=t.state.value;r.length>=o?t.notify(e):n.length>r.length&&t.notify(a(a({},e),{},{target:a(a({},e.target),{},{value:""})}))}))})),d(b(t),"onKeyDown",(function(e){"Enter"===e.key&&t.forceNotify(e);var n=t.props.onKeyDown;n&&(e.persist(),n(e))})),d(b(t),"onBlur",(function(e){t.forceNotify(e);var n=t.props.onBlur;n&&(e.persist(),n(e))})),d(b(t),"createNotifier",(function(e){if(e<0)t.notify=function(){return null};else if(0===e)t.notify=t.doNotify;else{var n=(0,r.default)((function(e){t.isDebouncing=!1,t.doNotify(e)}),e);t.notify=function(e){t.isDebouncing=!0,n(e)},t.flush=function(){return n.flush()},t.cancel=function(){t.isDebouncing=!1,n.cancel()}}})),d(b(t),"doNotify",(function(){var e=t.props.onChange;e.apply(void 0,arguments)})),d(b(t),"forceNotify",(function(e){var n=t.props.debounceTimeout;if(t.isDebouncing||!(n>0)){t.cancel&&t.cancel();var o=t.state.value,r=t.props.minLength;o.length>=r?t.doNotify(e):t.doNotify(a(a({},e),{},{target:a(a({},e.target),{},{value:o})}))}})),t.isDebouncing=!1,t.state={value:"undefined"===typeof e.value||null===e.value?"":e.value};var n=t.props.debounceTimeout;return t.createNotifier(n),t}return t=f,(n=[{key:"componentDidUpdate",value:function(e){if(!this.isDebouncing){var t=this.props,n=t.value,o=t.debounceTimeout,r=e.debounceTimeout,i=e.value,u=this.state.value;"undefined"!==typeof n&&i!==n&&u!==n&&this.setState({value:n}),o!==r&&this.createNotifier(o)}}},{key:"componentWillUnmount",value:function(){this.flush&&this.flush()}},{key:"render",value:function(){var e,t,n=this.props,r=n.element,i=(n.onChange,n.value,n.minLength,n.debounceTimeout,n.forceNotifyByEnter),u=n.forceNotifyOnBlur,f=n.onKeyDown,l=n.onBlur,s=n.inputRef,p=c(n,["element","onChange","value","minLength","debounceTimeout","forceNotifyByEnter","forceNotifyOnBlur","onKeyDown","onBlur","inputRef"]),y=this.state.value;e=i?{onKeyDown:this.onKeyDown}:f?{onKeyDown:f}:{},t=u?{onBlur:this.onBlur}:l?{onBlur:l}:{};var b=s?{ref:s}:{};return o.default.createElement(r,a(a(a(a({},p),{},{onChange:this.onChange,value:y},e),t),b))}}])&&l(t.prototype,n),i&&l(t,i),f}(o.default.PureComponent);t.DebounceInput=h,d(h,"defaultProps",{element:"input",type:"text",onKeyDown:void 0,onBlur:void 0,value:void 0,minLength:0,debounceTimeout:100,forceNotifyByEnter:!0,forceNotifyOnBlur:!0,inputRef:void 0})},40:function(e,t,n){(function(t){var n=/^\s+|\s+$/g,o=/^[-+]0x[0-9a-f]+$/i,r=/^0b[01]+$/i,i=/^0o[0-7]+$/i,u=parseInt,c="object"==typeof t&&t&&t.Object===Object&&t,f="object"==typeof self&&self&&self.Object===Object&&self,a=c||f||Function("return this")(),l=Object.prototype.toString,s=Math.max,p=Math.min,y=function(){return a.Date.now()};function b(e){var t=typeof e;return!!e&&("object"==t||"function"==t)}function v(e){if("number"==typeof e)return e;if(function(e){return"symbol"==typeof e||function(e){return!!e&&"object"==typeof e}(e)&&"[object Symbol]"==l.call(e)}(e))return NaN;if(b(e)){var t="function"==typeof e.valueOf?e.valueOf():e;e=b(t)?t+"":t}if("string"!=typeof e)return 0===e?e:+e;e=e.replace(n,"");var c=r.test(e);return c||i.test(e)?u(e.slice(2),c?2:8):o.test(e)?NaN:+e}e.exports=function(e,t,n){var o,r,i,u,c,f,a=0,l=!1,d=!1,h=!0;if("function"!=typeof e)throw new TypeError("Expected a function");function m(t){var n=o,i=r;return o=r=void 0,a=t,u=e.apply(i,n)}function g(e){return a=e,c=setTimeout(j,t),l?m(e):u}function O(e){var n=e-f;return void 0===f||n>=t||n<0||d&&e-a>=i}function j(){var e=y();if(O(e))return w(e);c=setTimeout(j,function(e){var n=t-(e-f);return d?p(n,i-(e-a)):n}(e))}function w(e){return c=void 0,h&&o?m(e):(o=r=void 0,u)}function D(){var e=y(),n=O(e);if(o=arguments,r=this,f=e,n){if(void 0===c)return g(f);if(d)return c=setTimeout(j,t),m(f)}return void 0===c&&(c=setTimeout(j,t)),u}return t=v(t)||0,b(n)&&(l=!!n.leading,i=(d="maxWait"in n)?s(v(n.maxWait)||0,t):i,h="trailing"in n?!!n.trailing:h),D.cancel=function(){void 0!==c&&clearTimeout(c),a=0,o=f=r=c=void 0},D.flush=function(){return void 0===c?u:w(y())},D}}).call(this,n(24))},43:function(e,t,n){"use strict";n.r(t);n(1);var o=n(38),r=n(0);t.default=function(e){var t=e.label,n=e.onChange,i=e.value,u=e.id,c=e.debounceTimeout,f=void 0===c?200:c;return Object(r.jsx)("div",{className:"form-inline justify-content-end","data-filter":!0,children:Object(r.jsxs)("div",{className:"form-group",children:[Object(r.jsx)("label",{htmlFor:"".concat(u,"FilterInput"),children:t}),Object(r.jsx)(o.DebounceInput,{style:{marginLeft:"0.75em",display:"inline-block",width:"250px"},id:"".concat(u,"FilterInput"),className:"form-control",onChange:function(e){return n(e)},value:i,debounceTimeout:f,"data-filter-input":!0})]})})}}}]);
+//# sourceMappingURL=3.20a0c265.chunk.js.map
