+++ docs/static/js/4.283b7797.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonp@mydobie/sort-table"]=this["webpackJsonp@mydobie/sort-table"]||[]).push([[4],{41:function(e,t,s){},45:function(e,t,s){"use strict";s.r(t);s(1),s(41);var a=s(0),n=function(e){var t=e.header,s=e.tableData;return Object(a.jsxs)(a.Fragment,{children:[Object(a.jsx)("dt",{children:t.name}),Object(a.jsx)("dd",{children:s[t.key]})]})};t.default=function(e){var t=e.headers,s=e.tableData,r=e.isResponsiveListAlwaysShow;return Object(a.jsx)("div",{"data-sort-responsive-list":!0,"data-sort-responsive-list-always-show":r,children:s.map((function(e){return Object(a.jsx)("dl",{children:t.map((function(t){return Object(a.jsx)(n,{header:t,tableData:e},t.key)}))},e.id)}))})}}}]);
+//# sourceMappingURL=4.283b7797.chunk.js.map
