+++ docs/static/js/5.23c74361.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonp@mydobie/sort-table"]=this["webpackJsonp@mydobie/sort-table"]||[]).push([[5],{37:function(e,n,a){},47:function(e,n,a){"use strict";a.r(n);var t=a(1),i=a(0),c=function(e){var n,a=e.active,t=e.disabled,c=e.pageNumber,s=e.onClick,o=e.isPrevious,r=e.isNext;return r?n="Next page":o&&(n="Previous page"),Object(i.jsx)("li",{className:"page-item ".concat(t?"disabled":""," ").concat(a?"active":""),"data-pagination-previous-button":o,"data-pagination-next-button":r,"data-pagination-button":!r&&!o||void 0,children:Object(i.jsxs)("button",{type:"button",className:"page-link ".concat(a?"active":""),"aria-current":a?"page":void 0,onClick:s?function(){return s(c)}:function(){},disabled:t,"aria-label":n,children:[o?Object(i.jsx)("span",{"aria-hidden":"true",children:"\xab"}):null,r?Object(i.jsx)("span",{"aria-hidden":"true",children:"\xbb"}):null,o||r?null:Object(i.jsxs)(i.Fragment,{children:[Object(i.jsx)("span",{className:"sr-only visually-hidden",children:"page "}),c]})]})})};a(37),n.default=function(e){var n=e.activePage,a=void 0===n?1:n,s=e.numberOfPages,o=e.onPageChange,r=e.id;Object(t.useEffect)((function(){a>s&&o(s),a<1&&o(1)}),[a,s,o]),s<1&&console.warn("Number of pagination pages is less than 1");if(s<=1)return null;return Object(i.jsxs)("div",{children:[Object(i.jsx)("div",{className:"text-right text-end","data-skip-pagination-wrapper":!0,children:Object(i.jsx)("a",{href:"#".concat(r,"EndPagination"),"data-skip-pagination-link":!0,children:"Skip pagination"})}),Object(i.jsx)("nav",{"aria-label":"Table pagination",children:Object(i.jsxs)("ul",{className:"pagination pagination-sm justify-content-end","data-pagination":!0,children:[Object(i.jsx)(c,{active:0===a,disabled:a<=1,onClick:function(){o(a-1)},isPrevious:!0}),s>10?function(){for(var e=[],n=1;n<=s;n+=1)e.push(Object(i.jsx)("option",{value:n,children:n},n));return Object(i.jsx)("li",{children:Object(i.jsx)("select",{className:"form-select form-select-sm custom-select custom-select-sm",style:{width:"6em"},onChange:function(e){return o(parseInt(e.target.value,10))},value:a,"aria-label":"Show page number","data-pagination-select":!0,children:e})})}():function(){for(var e=[],n=function(n){e.push(Object(i.jsx)(c,{pageNumber:n,active:a===n,onClick:function(){o(n)}},n))},t=1;t<=s;t+=1)n(t);return e}(),Object(i.jsx)(c,{active:a===s+1,disabled:a>=s,onClick:function(){o(a+1)},isNext:!0})]})}),Object(i.jsx)("div",{id:"".concat(r,"EndPagination")})]})}}}]);
+//# sourceMappingURL=5.23c74361.chunk.js.map
