+++ docs/static/js/main.20033631.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonp@mydobie/sort-table"]=this["webpackJsonp@mydobie/sort-table"]||[]).push([[0],{19:function(e,t,a){},35:function(e,t,a){},36:function(e,t,a){"use strict";a.r(t);var n=a(1),s=a.n(n),i=a(20),l=a.n(i),o=a(9),r=(a(29),"/"),c=a(0),d=function(){return Object(c.jsx)("nav",{children:Object(c.jsx)("ul",{className:"nav",children:Object(c.jsx)("li",{className:"nav-item",children:Object(c.jsx)(o.b,{activeClassName:"active",className:"nav-link",to:r,children:"Home"})})})})},u=a(2),h=a(8),j=function(e){var t=e.sortAsc,a=void 0===t||t,n=e.color||"#000000",s=e.size||"20px",i=e.type||"size",l={width:s,height:s,display:"inline-block",paddingRight:"5px"},o=Object(c.jsx)("path",{"data-icontype":"alphaAscending",fill:n,d:"M176 352h-48V48a16 16 0 0 0-16-16H80a16 16 0 0 0-16 16v304H16c-14.19 0-21.36 17.24-11.29 27.31l80 96a16 16 0 0 0 22.62 0l80-96C197.35 369.26 190.22 352 176 352zm240-64H288a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h56l-61.26 70.45A32 32 0 0 0 272 446.37V464a16 16 0 0 0 16 16h128a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16h-56l61.26-70.45A32 32 0 0 0 432 321.63V304a16 16 0 0 0-16-16zm31.06-85.38l-59.27-160A16 16 0 0 0 372.72 32h-41.44a16 16 0 0 0-15.07 10.62l-59.27 160A16 16 0 0 0 272 224h24.83a16 16 0 0 0 15.23-11.08l4.42-12.92h71l4.41 12.92A16 16 0 0 0 407.16 224H432a16 16 0 0 0 15.06-21.38zM335.61 144L352 96l16.39 48z"}),r=Object(c.jsx)("path",{"data-icontype":"defaultAscending",fill:n,d:"M240 96h64a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16h-64a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16zm0 128h128a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16zm256 192H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h256a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16zm-256-64h192a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16zm-64 0h-48V48a16 16 0 0 0-16-16H80a16 16 0 0 0-16 16v304H16c-14.19 0-21.37 17.24-11.29 27.31l80 96a16 16 0 0 0 22.62 0l80-96C197.35 369.26 190.22 352 176 352z"}),d=Object(c.jsx)("path",{"data-icontype":"alphaDescending",fill:n,d:"M16 160h48v304a16 16 0 0 0 16 16h32a16 16 0 0 0 16-16V160h48c14.21 0 21.38-17.24 11.31-27.31l-80-96a16 16 0 0 0-22.62 0l-80 96C-5.35 142.74 1.78 160 16 160zm272 64h128a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16h-56l61.26-70.45A32 32 0 0 0 432 65.63V48a16 16 0 0 0-16-16H288a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h56l-61.26 70.45A32 32 0 0 0 272 190.37V208a16 16 0 0 0 16 16zm159.06 234.62l-59.27-160A16 16 0 0 0 372.72 288h-41.44a16 16 0 0 0-15.07 10.62l-59.27 160A16 16 0 0 0 272 480h24.83a16 16 0 0 0 15.23-11.08l4.42-12.92h71l4.41 12.92A16 16 0 0 0 407.16 480H432a16 16 0 0 0 15.06-21.38zM335.61 400L352 352l16.39 48z"}),u=Object(c.jsx)("path",{"data-icontype":"defaultDescending",fill:n,d:"M304 416h-64a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h64a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16zM16 160h48v304a16 16 0 0 0 16 16h32a16 16 0 0 0 16-16V160h48c14.21 0 21.38-17.24 11.31-27.31l-80-96a16 16 0 0 0-22.62 0l-80 96C-5.35 142.74 1.77 160 16 160zm416 0H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h192a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16zm-64 128H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h128a16 16 0 0 0 16-16v-32a16 16 0 0 0-16-16zM496 32H240a16 16 0 0 0-16 16v32a16 16 0 0 0 16 16h256a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16z"}),h=Object(c.jsx)("path",{"data-icontype":"sortable",fill:n,d:"M176 352h-48V48a16 16 0 0 0-16-16H80a16 16 0 0 0-16 16v304H16c-14.19 0-21.36 17.24-11.29 27.31l80 96a16 16 0 0 0 22.62 0l80-96C197.35 369.26 190.22 352 176 352zm203.29-219.31l-80-96a16 16 0 0 0-22.62 0l-80 96C186.65 142.74 193.78 160 208 160h48v304a16 16 0 0 0 16 16h32a16 16 0 0 0 16-16V160h48c14.19 0 21.36-17.24 11.29-27.31z"}),j=Object(c.jsx)(c.Fragment,{});switch(i){case"alpha":j=!0===a?o:d;break;case"sortable":j=h;break;default:j=!0===a?r:u}return Object(c.jsx)("svg",{role:"img",xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 448 512",style:l,"aria-hidden":"true",children:j})},b=(a(35),function(e){var t=e.totalEntries,a=e.startRow,n=e.endRow,s=e.filteredTotal,i=e.sortColumn,l=e.sortDirection,o=n,r=null!==s&&void 0!==s?s:t;n&&s&&n>s&&(o=s);var d=a&&o&&0!==r&&(1!==a||o!==r)?"".concat(a," - ").concat(o," of "):"";return d+="".concat(r," "),d+=1===r?"entry":"entries",d+=void 0!==s&&s<t?" (filtered from ".concat(t,")"):"",d+=".",d+=i&&l?" ".concat(i," sorted ").concat(l,"."):"",Object(c.jsx)("span",{children:d})}),m=function(e){var t=e.headers,a=e.selected,n=void 0===a?null:a,s=e.onChange,i=void 0===s?function(){}:s,l=e.sortAscending,o=void 0===l?null:l,r=e.id,d=void 0===r?"sortDropDown":r;return Object(c.jsxs)("div",{className:"row","data-sort-dropdown":!0,"data-testid":"sortDropDownWrapper",children:[Object(c.jsx)("div",{className:"col-auto",style:{paddingRight:0},children:Object(c.jsx)("label",{htmlFor:"".concat(d,"SortDropDown"),children:"Sort:"})}),Object(c.jsx)("div",{className:"col-auto",style:{paddingRight:0},children:Object(c.jsxs)("select",{id:"".concat(d,"SortDropDown"),value:n||"",className:"form-control form-select",onChange:function(e){return i(e.target.value,!0)},"data-testid":"sortDropDownColumn",children:[""===n?Object(c.jsx)("option",{value:"",children:"Select a sort"}):null,t.map((function(e){return e.noSort?null:Object(c.jsx)("option",{value:e.sortKey||e.key||"",children:e.name},e.key)}))]})}),n?Object(c.jsx)("div",{className:"col-auto",style:{paddingLeft:0},children:Object(c.jsx)("button",{type:"button",onClick:function(){i(n,!o)},style:{border:"none",background:"none",width:"100%",textAlign:"left",padding:"5px 10px"},"aria-label":"Sort order","data-testid":"sortDropDownButton","data-sortdirection":o?"asc":"desc",children:Object(c.jsx)(j,{sortAsc:o||!1})})}):null]})},v=a(17),p=a(16),y=a(23),x=function(e){var t=e.rows,a=e.sortCol,n=e.sortAscending,s=e.onSort,i=void 0===s?function(){}:s,l=Object(p.a)(t).sort((function(e,t){return e[a]===t[a]?0:void 0===e[a]?!1===n?-1:1:void 0===t[a]||e[a]<t[a]?!0===n?-1:1:!1===n?-1:1})).map((function(e,t){return Object(v.a)(Object(v.a)({},e),{},{rowindex:t+2})}));return i(l),l},f=function(e){var t=e.rows,a=e.filterValue,n=void 0===a?"":a,s=e.caseSensitiveFilter,i=void 0!==s&&s,l=e.headers,o=e.useFuzzySearch,r=void 0===o||o,c=e.maxFuzzyDistance,d=!0===i?n:n.toLowerCase(),u=Object(p.a)(t);return u.forEach((function(e,t){u[t].hide=!0,l.forEach((function(a){if((void 0===a.noFilter||!1===a.noFilter)&&void 0!==e[a.key]){var n=e[a.key].toString();n=!0===i?n:n.toLowerCase(),(r&&Object(y.distance)(n,d)<=c||n.includes(d))&&(u[t].hide=!1)}}))})),t},O=s.a.lazy((function(){return a.e(5).then(a.bind(null,47))})),g=s.a.lazy((function(){return a.e(3).then(a.bind(null,43))})),w=s.a.lazy((function(){return a.e(7).then(a.bind(null,44))})),S=s.a.lazy((function(){return a.e(4).then(a.bind(null,45))})),C=s.a.lazy((function(){return a.e(6).then(a.bind(null,46))})),k=function(e){var t,a,n=e.headers,i=e.tableData,l=e.allDataFilteredMessage,o=e.caption,r=e.initialFilter,d=void 0===r?"":r,u=e.initialPage,v=e.initialRowsDisplayed,p=e.initialSort,y=e.initialSortDsc,k=e.isLoadingMessage,z=e.noDataMessage,F=e.viewSteps,N=e.caseSensitiveFilter,D=e.dangerouslySetInnerHTML,A=e.defaultToAll,R=e.emptyCellClassName,H=e.headerClassName,T=e.id,M=e.isLoading,V=e.isResponsive,L=e.isResponsiveList,P=e.showFilter,$=e.showPagination,E=e.sortedCellClass,K=e.tableClassName,B=e.useFuzzySearch,I=e.onChange,W=void 0===I?function(){}:I,G=e.debounceTimeout,J=e.isResponsiveListAlwaysShow,_=e.maxFuzzyDistance,q=void 0===_?3:_,Y=A||!F?null:F[0];v&&F&&(Y=F.includes(v)?v:Y);var Q=null!==T&&void 0!==T?T:"sortTable",U=s.a.useState(i),X=Object(h.a)(U,2),Z=X[0],ee=X[1],te=s.a.useState(""),ae=Object(h.a)(te,2),ne=ae[0],se=ae[1],ie=s.a.useState(!0),le=Object(h.a)(ie,2),oe=le[0],re=le[1],ce=s.a.useState(""),de=Object(h.a)(ce,2),ue=de[0],he=de[1],je=s.a.useState(Y),be=Object(h.a)(je,2),me=be[0],ve=be[1],pe=s.a.useState(null!==u&&void 0!==u?u:1),ye=Object(h.a)(pe,2),xe=ye[0],fe=ye[1],Oe=s.a.useCallback((function(){return Z.filter((function(e){return!e.hide}))}),[Z]),ge=Math.ceil(me&&me>0&&Oe().length>0?Oe().length/me:1),we=function(e,t){var a=arguments.length>2&&void 0!==arguments[2]?arguments[2]:Z,n=x({rows:a,sortCol:e,sortAscending:t});ee(n),se(e),re(t)},Se=me?(xe-1)*me:1,Ce=null!==z&&void 0!==z?z:Object(c.jsx)("p",{"data-sort-no-data-message":!0,children:"No data is available"}),ke=null!==l&&void 0!==l?l:Object(c.jsx)("p",{"data-sort-all-data-filtered":!0,children:"No data meets filtering criteria"});s.a.useEffect((function(){var e=p?n.find((function(e){return e.key===p})):void 0;if(e){var t,a=null!==(t=e.sortKey)&&void 0!==t?t:e.key;x({rows:Z,sortCol:a,sortAscending:!y,onSort:function(e){if(se(a),re(!y),d&&""!==d){var t=f({rows:e,filterValue:d,caseSensitiveFilter:N,headers:n,useFuzzySearch:B,maxFuzzyDistance:q});ee(t),he(d)}else ee(e)}})}}),[]),s.a.useEffect((function(){W({sortedColumn:ne,sortedAscending:oe,rowsShown:me,filter:ue,page:xe,pages:ge,totalFiltered:Oe().length})}),[xe,ue,Oe,me,ge,W,oe,ne]);var ze=function(e){if(void 0!==e.noSort&&!0===e.noSort)return e.name;var t=e.key===ne||e.sortKey===ne?e.type:"sortable",a="sortable"===t?"#ccc":void 0;return Object(c.jsxs)("button",{type:"button",onClick:function(){var t,a=null!==(t=e.sortKey)&&void 0!==t?t:e.key;we(a,a!==ne||!oe)},style:{border:"none",padding:"none",background:"none",width:"100%",textAlign:"left"},children:[Object(c.jsx)(j,{sortAsc:oe,type:t,color:a}),e.name]})},Fe=function(e){if(e===ne)return oe?"ascending":"descending"},Ne=Object(c.jsx)("tr",{"aria-rowindex":$?1:void 0,role:V&&!L?"row":void 0,children:n.map((function(e){return Object(c.jsx)("th",{scope:"col","aria-sort":Fe(e.key),style:e.style,className:"".concat(e.className," ").concat(e.key===ne||e.sortKey===ne?E:""),role:V&&!L?"columnheader":void 0,children:ze(e)},e.key)}))}),De=function(){return $&&me?Oe().slice(Se,Se+me):Oe()};return 0===Z.length?Object(c.jsx)(c.Fragment,{children:Ce}):Object(c.jsx)("div",{className:"container-fluid",children:Object(c.jsxs)(s.a.Suspense,{fallback:Object(c.jsx)(c.Fragment,{}),children:[V&&!L?Object(c.jsx)(C,{}):null,Object(c.jsxs)("div",{className:"row",style:{marginBottom:$||P?"15px":"0"},children:[$?Object(c.jsx)("div",{className:"col-sm",children:Object(c.jsxs)("div",{"data-sort-number-of-inputs":!0,children:["Show",Object(c.jsxs)("select",{className:"form-control form-select","aria-label":"Number of items shown",value:me||"",style:{width:"75px",display:"inline-block",marginLeft:"0.75em",marginRight:"0.75em"},onChange:function(e){var t;fe(1),ve(null!==(t=parseInt(e.target.value,10))&&void 0!==t?t:null)},children:[null===F||void 0===F?void 0:F.map((function(e){return Object(c.jsx)("option",{value:e,children:e},e)})),Object(c.jsx)("option",{value:"",children:"All"})]}),"results"]})}):null,P?Object(c.jsx)("div",{className:"col-sm",style:{textAlign:"right"},children:Object(c.jsx)(g,{value:ue,onChange:function(e){var t;t=e.target.value,ee(f({rows:Z,filterValue:t,caseSensitiveFilter:N,headers:n,useFuzzySearch:B,maxFuzzyDistance:q})),he(t)},label:"Filter",id:Q,debounceTimeout:G})}):null]}),Object(c.jsxs)("div",{className:"row",children:[V||L?Object(c.jsx)(m,{headers:n,selected:ne,sortAscending:oe,onChange:function(e,t){we(e,t)}}):null,Object(c.jsxs)("table",{className:"table ".concat(K),id:Q,"aria-describedby":"".concat(Q,"RowsShownSummary"),"aria-rowcount":$||P?Z.length+1:void 0,"data-sort-responsive":null!==(t=V&&!L)&&void 0!==t?t:void 0,"data-sort-responsive-has-list":null!==L&&void 0!==L?L:void 0,"data-sort-responsive-has-list-always-hide":null!==J&&void 0!==J?J:void 0,role:V&&!L?"table":void 0,children:[o?Object(c.jsx)("caption",{children:o}):null,Object(c.jsx)("thead",{className:H,role:V&&!L?"rowgroup":void 0,children:Ne}),M?null:Object(c.jsx)("tbody",{role:V&&!L?"rowgroup":void 0,"data-testid":"sortTableBody",children:De().map((function(e){return t=e,Object(c.jsx)("tr",{"aria-rowindex":$?t.rowindex:void 0,role:V&&!L?"row":void 0,children:n.map((function(e){var a=D?Object(c.jsx)("span",{dangerouslySetInnerHTML:{__html:t[e.key]}}):t[e.key],n="".concat(a?"":R," ").concat(e.key===ne||e.sortKey===ne?E:"");return e.rowheader?Object(c.jsx)("th",{scope:"row","data-sorttable-data-cell":!0,role:V&&!L?"rowheader":void 0,className:n,children:a},"".concat(t.id).concat(e.key)):Object(c.jsxs)("td",{"data-sorttable-data-cell":!0,role:V&&!L?"cell":void 0,className:n,children:[V&&!L?Object(c.jsx)("span",{"aria-hidden":!0,"data-responsive-header":!0,children:e.name}):null,a]},"".concat(t.id).concat(e.key))}))},t.id);var t}))})]}),L?Object(c.jsx)(S,{headers:n,tableData:De(),isResponsiveListAlwaysShow:J}):null,-1===Z.findIndex((function(e){return!e.hide}))?ke:null,M&&!k?Object(c.jsx)(w,{}):null,M&&k?k:null]}),Object(c.jsxs)("div",{className:"row",children:[Object(c.jsx)("div",{className:"col-sm",id:"".concat(Q,"RowsShownSummary"),"data-pagination-summary":!0,children:M?null:Object(c.jsx)(b,{totalEntries:i.length,startRow:me?Se+1:0,endRow:me?Se+me:Oe().length,filteredTotal:Oe().length,sortColumn:null===(a=n.find((function(e){return e.key===ne||e.sortKey===ne})))||void 0===a?void 0:a.name,sortDirection:oe?"ascending":"descending"})}),$?Object(c.jsx)("div",{className:"col-sm",children:Object(c.jsx)(O,{numberOfPages:ge,activePage:xe,id:Q,onPageChange:function(e){fe(e)}})}):null]})]})})};a(19);var z=function(){var e=[{name:"Product Name",key:"name",type:"alpha",rowheader:!0},{name:"Price",key:"price",className:"myCustomPriceClass"},{name:"Stock",key:"stock",type:"number",noFilter:!0},{name:"Sale day",key:"day",type:"size",sortKey:"saledaynum"},{name:"Link",key:"url",noSort:!0,noFilter:!0,style:{color:"purple"}}],t=[{id:1,name:"Cheese",price:"$4.90",stock:20,day:"Friday",saledaynum:5,url:""},{id:2,name:"Milk",price:"$1.90",stock:4,day:"Monday",saledaynum:1},{id:3,name:"Yoghurt",price:"$2.40",stock:12,day:"Thursday",saledaynum:4},{id:4,name:"Heavy Cream",price:"$3.90",stock:9,day:"Tuesday",saledaynum:2},{id:5,name:"Butter",price:"$0.90",stock:99,day:"Wednesday",saledaynum:3},{id:6,name:"Sour Cream",price:"$2.90",stock:86,day:"Saturday",saledaynum:6},{id:7,name:"Fancy French Cheese",price:"$99.0",stock:15,url:Object(c.jsx)("a",{href:"https://google.com",children:"hello google"}),day:"Sunday",saledaynum:0},{id:8,name:"Eggs",price:"$1.90",stock:86},{id:9,name:"Ricotta Cheese",price:"$3.99",stock:6,day:"Saturday",saledaynum:6},{id:10,name:"Frozen Custard",price:"$5.50",stock:11},{id:11,name:"Whey Powder",price:"$15.90",stock:8,day:"Wednesday",saledaynum:3}];return Object(c.jsxs)("div",{className:"container",children:[Object(c.jsx)("h1",{children:"Sample Filtering and Sorting Table"}),Object(c.jsxs)("p",{children:["This is the demo page for a React sort table component. See the"," ",Object(c.jsx)("a",{href:"https://github.com/mydobie/sortTable",children:"Github page"})," for more information."]}),Object(c.jsx)("hr",{}),Object(c.jsx)("h2",{children:"Responsive using modified CSS"}),Object(c.jsx)("p",{children:"This table has sorting, filter, and pagination enable. To assist users to know what column is sorted, the sorted cells have a custom CSS applied. On screens less than 700px, custom CSS will be applied to display the table to appear as if it is a list."}),Object(c.jsx)(k,{id:"sampleTable1",tableData:t,headers:e,initialSort:"name",initialSortDsc:!0,showFilter:!0,showPagination:!0,viewSteps:[1,2,4,50],caption:"Store inventory",tableClassName:"table-hover table-sm",defaultToAll:!0,isResponsive:!0,emptyCellClassName:"emptyCell",sortedCellClass:"sortedCellClass",useFuzzySearch:!0,initialPage:2,initialFilter:"cheese",initialRowsDisplayed:2,allDataFilteredMessage:"My custom all filtered message",onChange:function(e){console.log(e)}}),Object(c.jsx)("hr",{}),Object(c.jsx)("h2",{children:"Responsive using definition list"}),Object(c.jsx)("p",{children:"On screens less than 600px, the table will be hidden and a definition list will be shown instead."}),Object(c.jsx)(k,{id:"sampleTable2",tableData:t,headers:e,initialSort:"month",caption:"Store inventory",tableClassName:"table-hover table-sm",isResponsiveList:!0}),Object(c.jsx)("hr",{}),Object(c.jsxs)("ul",{children:[Object(c.jsxs)("li",{children:[Object(c.jsx)("strong",{children:"Project Name: "}),"@mydobie/sort-table"]}),Object(c.jsxs)("li",{children:[Object(c.jsx)("strong",{children:"Project Version: "}),"3.2.0"]}),Object(c.jsxs)("li",{children:[Object(c.jsx)("strong",{children:"Git Commit: "}),"40922deb"]}),Object(c.jsxs)("li",{children:[Object(c.jsx)("strong",{children:"React Version: "}),s.a.version]})]})]})},F=function(){return Object(c.jsxs)(c.Fragment,{children:[Object(c.jsx)("h1",{children:"Page not found"}),Object(c.jsx)("p",{children:"The page you requested could not be found."})]})},N=function(){return Object(c.jsx)(c.Fragment,{children:Object(c.jsxs)(u.d,{children:[Object(c.jsx)(u.b,{path:r,exact:!0,children:Object(c.jsx)(z,{})}),Object(c.jsx)(u.b,{path:"/home",exact:!0,children:Object(c.jsx)(u.a,{to:r})}),Object(c.jsx)(u.b,{path:"/",children:Object(c.jsx)(F,{})})]})})},D=function(){return Object(c.jsx)(c.Fragment,{children:Object(c.jsxs)(o.a,{basename:"",children:[Object(c.jsx)(d,{}),Object(c.jsx)("main",{children:Object(c.jsx)(N,{})})]})})},A=Object(c.jsx)(D,{});l.a.render(A,document.getElementById("root"))}},[[36,1,2]]]);
+//# sourceMappingURL=main.20033631.chunk.js.map
