+++ package.json
@@ -23,38 +23,40 @@
     "react-debounce-input": "^3.2.5"
   },
   "devDependencies": {
+    "bootstrap": "^5.1.3",
     "react": "16.8.0 - 17.999.999",
     "react-dom": "16.8.0 - 17.999.999",
-    "@babel/preset-react": "^7.12.10",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.2",
-    "@types/jest-axe": "^3.5.3",
-    "@types/lodash": "^4.14.175",
-    "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.9",
+    "@testing-library/react-hooks": "^7.0.1",
+    "@testing-library/user-event": "^13.4.1",
+    "@types/jest-axe": "^3.5.1",
     "@types/react-router-dom": "^5.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "bootstrap": "5.1.3",
+    "@types/styled-components": "5.1.15",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-config-airbnb-typescript": "^14.0.1",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-watch": "^7.0.0",
-    "husky": "^7.0.2",
+    "husky": "^7.0.0",
     "jest-axe": "^5.0.1",
-    "lint-staged": "^11.2.2",
-    "prettier": "2.4.1",
+    "lint-staged": "^11.2.3",
+    "onchange": "^7.1.0",
+    "prettier": "^2.3.2",
     "react-scripts": "^4.0.3",
     "sass": "1.39.2",
+    "stylelint": "^13.12.0",
+    "stylelint-config-prettier": "^9.0.3",
     "typescript": "^4.4.4",
-    "uglify-js": "^3.14.2"
+    "uglify-js": "^3.12.3"
   },
   "resolutions": {
     "glob-parent": "5.1.2",
@@ -79,7 +81,7 @@
     "lint:fix": "npm run prettier:fix:all && npm run lint:js:fix ",
     "lint:watch": "onchange 'src/**/*.*' -- npm run lint:fix",
     "----------------------------------------------------------3": "",
-    "lint:js": "esw --ext .jsx,.js --color ",
+    "lint:js": "tsc --noEmit && esw --ext .js,.jsx,.ts,.tsx --color ",
     "lint:js:fix": "npm run lint:js -- --fix",
     "prettier": "prettier  --ignore-path ./.eslintignore --check",
     "prettier:all": "npm run prettier '**/*.+(js|jsx|html|css|scss|json|md)'",
@@ -93,83 +95,67 @@
     "npmAudit": "npm audit --production --audit-level=high --registry=https://registry.npmjs.org",
     "buildPackage": "npm run d && npm run npmAudit && npm run test && bash  utils/buildPackage.sh"
   },
-  "babel": {
-    "presets": [
-      "react",
-      "env"
-    ]
-  },
   "prettier": {
     "singleQuote": true,
     "jsxSingleQuote": true,
     "printWidth": 80
   },
   "eslintConfig": {
-    "settings": {
-      "import/resolver": {
-        "typescript": {}
-      }
-    },
-    "extends": [
-      "airbnb",
-      "airbnb/hooks",
-      "react-app",
-      "react-app/jest",
-      "prettier"
+    "root": true,
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "react",
+      "@typescript-eslint",
+      "prettier",
+      "import"
     ],
-    "globals": {
-      "styles": true
-    },
+    "extends": [
+      "airbnb-typescript",
+      "airbnb/hooks",
+      "eslint:recommended",
+      "plugin:react/recommended",
+      "plugin:@typescript-eslint/recommended",
+      "prettier",
+      "plugin:prettier/recommended"
+    ],
     "env": {
       "browser": true,
-      "commonjs": true,
+      "jasmine": true,
       "jest": true,
-      "es6": true,
-      "jquery": true
+      "node": true
     },
-    "parser": "babel-eslint",
-    "plugins": [
-      "prettier",
-      "eslint-plugin-react"
+    "parserOptions": {
+      "project": "./tsconfig.json"
+    },
+    "overrides": [
+      {
+        "files": [
+          "*.ts",
+          "*.tsx"
+        ],
+        "parserOptions": {
+          "project": [
+            "./tsconfig.json"
+          ]
+        }
+      }
     ],
     "rules": {
       "react/static-property-placement": "off",
+      "react/require-default-props": "off",
+      "react/prop-types": "off",
       "prettier/prettier": "error",
+      "@typescript-eslint/no-unused-vars": [
+        "error"
+      ],
       "no-var": "error",
       "curly": "error",
-      "one-var": "error",
       "no-console": "error",
       "no-unused-vars": "error",
       "arrow-body-style": [
         "error",
         "as-needed"
-      ],
-      "react/jsx-filename-extension": [
-        2,
-        {
-          "extensions": [
-            ".js",
-            ".jsx",
-            ".ts",
-            ".tsx"
-          ]
-        }
-      ],
-      "import/extensions": [
-        "error",
-        "ignorePackages",
-        {
-          "js": "never",
-          "jsx": "never",
-          "ts": "never",
-          "tsx": "never"
-        }
       ]
-    }
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged && npm run npmAudit"
     }
   },
   "jest": {
