+++ src/AppRouteNames.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/no-anonymous-default-export
+export default {
+  HOME: '/',
+};
