+++ src/index.tsx
@@ -1,13 +1,7 @@
-// NOTE:  There normally isn't a need to change this file
-
-import React from 'react';
+import React, { ReactElement } from 'react';
 import ReactDOM from 'react-dom';
-
 import App from './App';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const RenderApp: ReactElement = <App />;
+
+ReactDOM.render(RenderApp, document.getElementById('root'));
